Work on Inode Cache
