+++ rustcryptfs-linux/src/lib.rs
@@ -1,4 +1,6 @@
 mod encrypted_filesystem;
+mod inode_cache;
+
 pub mod error;
 
 pub use encrypted_filesystem::EncryptedFs;
